ref(onboarding): Split php onboarding docs
