+++ static/app/gettingStartedDocs/php/php/index.tsx
@@ -1,0 +1,23 @@
+import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {
+  feedbackOnboardingJsLoader,
+  replayOnboardingJsLoader,
+} from 'sentry/gettingStartedDocs/javascript/jsLoader/jsLoader';
+
+import {crashReport} from './crashReport';
+import {logs} from './logs';
+import {onboarding} from './onboarding';
+import {performance} from './performance';
+import {profiling} from './profiling';
+
+const docs: Docs = {
+  onboarding,
+  replayOnboardingJsLoader,
+  performanceOnboarding: performance,
+  profilingOnboarding: profiling,
+  crashReportOnboarding: crashReport,
+  feedbackOnboardingJsLoader,
+  logsOnboarding: logs,
+};
+
+export default docs;
