+++ static/app/gettingStartedDocs/php/symfony/index.tsx
@@ -1,0 +1,19 @@
+import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {
+  feedbackOnboardingJsLoader,
+  replayOnboardingJsLoader,
+} from 'sentry/gettingStartedDocs/javascript/jsLoader/jsLoader';
+
+import {crashReport} from './crashReport';
+import {onboarding} from './onboarding';
+import {profiling} from './profiling';
+
+const docs: Docs = {
+  onboarding,
+  replayOnboardingJsLoader,
+  profilingOnboarding: profiling,
+  crashReportOnboarding: crashReport,
+  feedbackOnboardingJsLoader,
+};
+
+export default docs;
